Update Pokedex list to move 10 and 25 (Up/Down, L/R)
Changes using left or right on the pokedex to move 10 instead of 7 (expected behavior for menus in the game), and adds the ability to use L or R to move 25 slots in the list- which should help navigate what can become a very large list quickly. Changes  all of these to work when held (as up and down do).